remove web test for now
